# app/http/module/qa/api_question_list.go
package qa

import (
	provider "bbs/app/provider/qa"
	"github.com/jader1992/gocore/framework/gin"
)

// QuestionList 获取问题列表
// @Summary 获取问题列表
// @Description 获取问题列表，包含作者信息，不包含回答
// @Accept json
// @Produce json
// @Tags qa
// @Param page query int false "列表页数"
// @Param size query int false "列表每页展示数量"
// @Success 200 {object} QuestionDTO "问题列表"
// @Router /question/list [get]
func (api *QApi) QuestionList(c *gin.Context) {
	qaService := c.MustMake(provider.QaKey).(provider.Service)

	page, _ := c.DefaultQueryInt("page", 1)
	size, _ := c.DefaultQueryInt("size", 10)

	start := (page - 1) * size
	pager := provider.Pager{
		Start: start,
		Size:  size,
	}

	questions, err := qaService.GetQuestions(c, &pager)
	if err != nil {
		c.ISetStatus(500).IText(err.Error())
		return
	}

	if err := qaService.QuestionsLoadAuthor(c, questions); err != nil {
		c.ISetStatus(500).IText(err.Error())
		return
	}

	questionsDTO := ConvertQuestionsToDTO(questions)

	c.ISetOkStatus().IJson(questionsDTO)
}